fix: 베스트상품 pageSize 동적으로 전달하기 위해 resize 함수 위치 변경하여 width 값 각 컴포넌트 전달로 수정
