+++ README.md
@@ -36,7 +36,7 @@
   - scrollToTop을 hook 폴더에서 관리하게 변경했습니다.
   - 현재 위치한 상단 메뉴에 파란글자가 나오게 하는 것을 누락하여 새로 적용했습니다.
   - 제품의 title을 description으로 설정했었는데 name으로 변경하였습니다.
-  - 
+  - (08.11) 베스트상품 pageSize를 하드코딩하였는데 동적 pageSize를 전달하도록 수정했습니다.
 
 ## 스크린샷
